Fix sample, add network building
Functionality is still incomplete - the connection from the output layer
to the input layer should output the target values instead of the actual
output values during training, to provide correct context.
Also, the output layer includes plan nodes in order to make the input
and output dimensions match; it shouldn't have any real negative impact,
but it's unnecessary and a bit messy.